+++ catalog-backend-module-gitea-github-compat/src/index.ts
@@ -1,4 +1,3 @@
 export { catalogModuleGiteaGithubCompat as default, catalogModuleGiteaGithubCompat } from './module';
 export { ACTION_MAP, TemplateCompatibilityProcessor } from './TemplateCompatibilityProcessor';
-export { GiteaGithubCompatEntityProvider } from './GiteaGithubCompatEntityProvider';
 export { readCompatibilityConfig } from './config';